fixes for relativeLUT

+++ apps/generateAccurateLUT.m
@@ -44,7 +44,7 @@ mirFile= [mirFolder filesep mirFile];
 % load the first image to get the self-reference region
 f = figure('Name', 'Please select a region of bare Si');
 im = double(imread(tifFile, 1));
-[~, selfRefRegion] = imcrop(im, double(median(im(:)))*[.8 1.2]);
+[~, selfRefRegion] = imcrop(im, median(im(:))*[.8 1.2]);
 pause(0.01); % so the window can close
 close(f);
